pyproject.toml
[project]
name = "open-recruiter"
version = "0.1.0"
description = "An autonomous AI agent for recruitment workflow automation"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
open-recruiter = "open_recruiter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/open_recruiter"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
